feat(TUP-29943):delete forever should also delete codejar folder https://jira.talendforge.org/browse/TUP-29943

+++ main/plugins/org.talend.core.repository/src/main/java/org/talend/core/repository/ui/actions/DeleteAction.java
@@ -1515,10 +1515,12 @@ public class DeleteAction extends AContextualAction {
                         .getFolder(ResourceUtils.getProject(ProjectManager.getInstance().getCurrentProject()),
                                 ERepositoryObjectType.getFolderName(currentJobNode.getObjectType()), true)
                         .getFolder(currentJobNode.getObject().getProperty().getLabel());
-                try {
-                    folder.delete(false, null);
-                } catch (CoreException e) {
-                    throw new PersistenceException(e);
+                if (folder != null) {
+                    try {
+                        folder.delete(false, null);
+                    } catch (CoreException e) {
+                        throw new PersistenceException(e);
+                    }
                 }
             }
         }
